Exercitiul1: added v2: changed while into for added space between the stars, without space at the end of line
Exercitiul2 - Find if there are enough positive numbers. User inputs the values and the number of positive numbers.
methods return boolean
methods accept an array and an int
validation is in place

highestNumber has as parameter an array

Exercitiul4
Created 3 independent methods to check if the given 3 int values are a triangle / an isosceles trian. / an echilateral trian.
the methods are boolean

 fokingcos
 made more attempts in solving the mathematics involved in Exercitiul4

 Main.java
 Updated according to the changes made in the classes